fine tunning bulk delete use case
