+++ app/api/core/application/BulkDeleteEntity.ts
@@ -1,15 +1,12 @@
 import { z } from 'zod';
 import { search } from 'api/search';
+import { MultiLanguageEntityDataSource } from 'api/entities.v2/contracts/MultiLanguageEntitiesDataSource';
+import { EntityDeletedEvent } from 'api/entities/events/EntityDeletedEvent';
 import { ArrayUtils } from 'api/common.v2/utils/Array';
 import { AbstractUseCase } from '../libs/UseCase';
-import { BatchDeleteEntityJob } from '../infrastructure/jobs/BatchDeleteEntityJob';
-import { EntityPermissionChecker, Specification } from '../domain/entity/EntityPermissionChecker';
 
 const InputSchema = z.object({
-  sharedIds: z
-    .array(z.string().trim())
-    .min(1, 'You must provide at least one sharedId for bulk deletion')
-    .max(1000, 'You must provide at most 1000 sharedIds for bulk deletion'),
+  sharedIds: z.array(z.string().trim()).min(1).max(100),
 });
 
 type Input = z.infer<typeof InputSchema>;
@@ -18,37 +15,33 @@ type Output = Input;
 
 type Deps = {
   search: typeof search;
-  entityPermissionChecker: EntityPermissionChecker;
+  entitiesDS: MultiLanguageEntityDataSource;
 };
 
 class BulkDeleteEntityUseCase extends AbstractUseCase<Input, Output, Deps> {
   static InputSchema = InputSchema;
 
   protected async executeAsync(input: Input): Promise<Output> {
-    const { sharedIds } = InputSchema.parse({
-      sharedIds: ArrayUtils.deduplicate(input.sharedIds, s => s),
-    });
+    const { sharedIds } = InputSchema.parse(input);
 
-    const grantedSharedIds = (
-      await this.deps.entityPermissionChecker.filterEntities(
-        sharedIds,
-        Specification.createDeleteSpecification(this.getActor())
-      )
-    ).getDataOrThrow();
+    const entityResult = await this.deps.entitiesDS.getAllBySharedId(sharedIds);
 
-    const chunks = ArrayUtils.splitInChunks(grantedSharedIds, 100);
+    if (entityResult.isError()) {
+      return input;
+    }
 
     await this.transactionManager.run(async () => {
-      await this.jobsDispatcher.dispatchMany(async dispatch =>
-        chunks.forEach(chunk => dispatch(BatchDeleteEntityJob, { sharedIds: chunk }))
-      );
-
-      await this.deps.search.bulkDeleteBySharedId(grantedSharedIds);
+      await this.deps.entitiesDS.bulkDelete(sharedIds);
+      await this.deps.search.bulkDeleteBySharedId(sharedIds);
     });
+
+    await ArrayUtils.sequentialFor(entityResult.getData(), async entity =>
+      this.eventBus.emit(EntityDeletedEvent.fromDomain(entity))
+    );
 
     return input;
   }
 }
 
 export { BulkDeleteEntityUseCase };
-export type { Input as BulkDeleteEntityInput, Output as BulkDeleteEntityOutput };
+export type { Input as BulkDeleteEntityUseCaseInput, Output as BulkDeleteEntityUseCaseOutput };
