+++ app/api/core/infrastructure/express/entity/BulkDeleteEntityController.ts
@@ -1,15 +1,17 @@
 import { AbstractController } from 'api/common.v2/infrastructure/AbstractController';
 import {
-  BulkDeleteEntityInput,
-  BulkDeleteEntityUseCase,
-} from 'api/core/application/BulkDeleteEntity';
+  RequestBulkDeleteEntityInput,
+  RequestBulkDeleteEntityUseCase,
+} from 'api/core/application/RequestBulkDeleteEntity';
 import { search } from 'api/search';
 import { DefaultDispatcher } from 'api/core/libs/queue/configuration/factories';
 import { tenants } from 'api/tenants';
 import entities from 'api/entities';
 import { TransactionManagerFactory } from '../../factories/TransactionManagerFactory';
+import { MongoEntityPermissionChecker } from '../../mongodb/entity/MongoEntityPermissionChecker';
+import { getConnection } from '../../mongodb/common/getConnectionForCurrentTenant';
 
-type RequestDto = BulkDeleteEntityInput;
+type RequestDto = RequestBulkDeleteEntityInput;
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 type ResponseDto = string;
@@ -19,8 +21,13 @@ class BulkDeleteEntityController extends AbstractController<RequestDto> {
     if (tenants.current()?.featureFlags?.v2BulkDeleteEntity) {
       const transactionManager = TransactionManagerFactory.default();
       const jobsDispatcher = DefaultDispatcher(this.tenantName);
+      const entityPermissionChecker = new MongoEntityPermissionChecker(
+        getConnection(),
+        transactionManager
+      );
 
-      const useCase = new BulkDeleteEntityUseCase({
+      const useCase = new RequestBulkDeleteEntityUseCase({
+        entityPermissionChecker,
         search,
         jobsDispatcher,
         transactionManager,
@@ -33,7 +40,7 @@ class BulkDeleteEntityController extends AbstractController<RequestDto> {
     }
 
     await entities.deleteMultiple(
-      BulkDeleteEntityUseCase.InputSchema.parse(this.request.body).sharedIds
+      RequestBulkDeleteEntityUseCase.InputSchema.parse(this.request.body).sharedIds
     );
 
     this.response.json('ok');
